requirements.txt
pandas>=2.1.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
pytest>=7.4.0
